+++ Changelog.md
@@ -8,6 +8,7 @@ Compiler Features:
 
 Bugfixes:
  * Code Generator: Fix internal compiler error when calling functions bound to calldata structs and arrays.
+ * Code Generator: Fix internal compiler error when passing zero to ``bytes.concat()``.
  * Type Checker: Fix internal error and prevent static calls to unimplemented modifiers.